requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
sentencepiece>=0.1.99
diffusers>=0.21.0
gradio>=4.0.0
pyyaml>=6.0
tqdm>=4.65.0
numpy>=1.24.0
pillow>=9.5.0
huggingface-hub>=0.16.0
datasets>=2.14.0
regex>=2023.0.0
accelerate>=0.20.0
pytest>=7.0.0
